requirements.txt
appnope==0.1.3
asgiref==3.8.1
astroid==2.0.4
asttokens==2.4.1
atomicwrites==1.1.5
attrs==18.1.0
beautifulsoup4==4.12.3
boto3==1.35.87
botocore==1.35.87
certifi==2024.12.14
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
cloudinary==1.41.0
comm==0.2.1
cryptography==44.0.0
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dj-rest-auth==7.0.0
Django==5.0.2
django-allauth==0.61.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.6.0
django-extensions==3.2.3
django-rest-passwordreset==1.5.0
django-seed==0.3.1
django-storages==1.14.4
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
executing==2.0.1
Faker==33.1.0
flake8==3.5.0
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.20.0
isort==4.3.4
jedi==0.19.1
jmespath==1.0.1
jupyter_client==8.6.0
jupyter_core==5.7.1
lazy-object-proxy==1.3.1
matplotlib-inline==0.1.6
mccabe==0.6.1
more-itertools==4.3.0
msgpack==1.1.0
nest-asyncio==1.6.0
oauthlib==3.2.2
packaging==23.2
parso==0.8.3
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.43
psutil==5.9.8
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.5.4
pycodestyle==2.3.1
pycparser==2.22
pyflakes==1.6.0
Pygments==2.17.2
PyJWT==2.10.1
pylint==2.1.1
pytest==7.4.4
pytest-django==4.9.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python3-openid==3.2.0
pyzmq==25.1.2
redis==5.2.0
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
setuptools==69.0.3
six==1.16.0
soupsieve==2.6
sqlparse==0.4.4
stack-data==0.6.3
toposort==1.10
tornado==6.4
traitlets==5.14.1
typing_extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
wrapt==1.10.11
